fix django-dsfr crashing with ImportError when optional dependency 'markdown' is absent
